+++ config/msegat.php
@@ -7,6 +7,11 @@ return [
     "username" => env("MSEGAT_USERNAME", ""),
     "sender" => env("MSEGAT_SENDER", ""),
     "unicode" => env("MSEGAT_UNICODE", "UTF8"),
-    "receiver" => env("MSEGAT_RECEIVER","phone"),
+    "receiver" => env("MSEGAT_RECEIVER", "phone"),
     "log" => env("MSEGAT_LOG", false),
+
+    "model" => [
+        "table_name" => "msegat_messages",
+        "allow_messages_log" => env("MSEGAT_MESSAGES_LOG", true)
+    ],
 ];