+++ src/MsegatChannel.php
@@ -2,6 +2,7 @@
 
 namespace BitcodeSa\Msegat;
 
+use BitcodeSa\Msegat\Models\Message;
 use Illuminate\Notifications\Notification;
 use Illuminate\Support\Facades\Cache;
 
@@ -35,8 +36,29 @@ class MsegatChannel
         if (is_string($message)) {
             $message = new MsegatMessage($message);
         }
+
         $this->msegat->setNotifiable($notifiable);
+
         $result = $this->msegat->sendMessage($recever, $message);
+
+        $this->createMessage($result, $notifiable, $message);
+
         return $result;
+    }
+
+    public function createMessage($response, $notifiable, $message)
+    {
+        if (config("msegat.model.allow_messages_log")) {
+            if ($response->successful()) {
+                Message::create([
+                    "phone" => $notifiable->{config("msegat.receiver")},
+                    "message" => $message->content,
+                    "messageable_type" => get_class($notifiable),
+                    "messageable_id" => $notifiable->{$notifiable->getKeyName()},
+                    "response_code" => $response->json("code", 0),
+                    "response_message" => $response->json("message", "No Response")
+                ]);
+            }
+        }
     }
 }